Fixes ViewLocationHolder.compareTo(), guarantees transitive property across comparisons.
I have an application that crashes whenever accessibility is on. The crash happens when the views on the activity are sorted; the sorting algorithm is a merge sort that tests as it merges, when it finds non-transitive comparisons (e.g., A>B, B>C, but !(A>C)), then it throws an exception. The current ViewLocationHolder.compareTo() does not guarantee transitivity.

My patch simplifies the comparison, and also changes it to guarantee that A>B and B>C will always mean A>C. I believe that it keeps the intent of the original sort order.

Change-Id: I375837241edb6b9b8ccb7a167731398f3af5a2b2